Fixes to get pass on windows.

+++ tests/test_data_dictionary.py
@@ -359,7 +359,7 @@ class TestDataDictionary:
         @model
         class TestDD:
             col1 = DD.def_parameter("COL1")
-            col1_mapping = DD.def_parameter("COL1", dtype_mapping={PandasDtype.Int: str})
+            col1_mapping = DD.def_parameter("COL1", dtype_mapping={PandasDtype.Int64: str})
             col2 = DD.def_parameter("COL2")
             col2_validator = DD.def_parameter("COL2", add_dtype_validator=True)
 
@@ -370,7 +370,7 @@ class TestDataDictionary:
         assert isinstance(
             TestDD(col1=1, col1_mapping=1, col2=1, col2_validator="x"), TestDD
         )
-        with pytest.raises(TypeError):  # fails due col2_validator should be str
+        with pytest.raises(TypeError):  # fails due col2_validator should be 
             TestDD(col1=1, col1_mapping=1, col2=1, col2_validator=1)
         with pytest.raises(ValueError):  # fails because col1 is 2
             TestDD(col1=2, col1_mapping=1, col2="x", col2_validator="x")
@@ -387,7 +387,7 @@ class TestDataDictionary:
         pytest.raises(AttributeError, DD.def_column, "zz")
         assert DD.def_column("COL1") == Column(
             name="COL1",
-            dtype="int",
+            dtype="int64",
             description="This is column 1.",
             validator=[equal_to(1)],
         )
